% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config_pki.R
\name{auto_config_pki}
\alias{auto_config_pki}
\title{Set httr config parameters}
\usage{
auto_config_pki()
}
\description{
This is a wrapper function that automatically overrides configuration settings used by the httr package to allow working with PKI-enabled web services.
}
\details{
This function should be called for automatic configuration. The PKI certificate and private key are extracted from a PKCS#12 file and used to define the following httr package config settings: cainfo, sslcert, and sslkey
}
\examples{
library(rpki)
auto_config_pki()
GET('https://your.pki.enabled.website/path/to/whatever')

}
