% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_download_file_pki.R
\name{configure_download_file_pki}
\alias{configure_download_file_pki}
\title{Set download.file parameters}
\usage{
configure_download_file_pki(mypki_file = NULL, pki_file = NULL,
  ca_file = NULL, password = NULL, overwrite = FALSE)
}
\arguments{
\item{mypki_file}{string: absolute file path to save mypki configuration file. Defaults to the home directory}

\item{pki_file}{string: absolute file path to a pki certificate}

\item{ca_file}{string: absolute file path to a Certificate Authority (CA) bundle}

\item{password}{string: passphrase used to encrypt/decrypt the private key of the PKI certificate}

\item{overwrite}{logical: overwrite a pre-existing mypki configuration file, if found. Defaults to FALSE}
}
\description{
Override default configuration settings used by download.file to allow for PKI-enabled use of install.packages().
}
\details{
Default mypki configuration settings will be used. The PKI certificate file must be in PKCS#12 format. The following download.file settings get modified: cacert, cert, key, pass.

If a mypki configuration file cannot be found, users are prompted for file paths to both a PKI certificate and a Certificate Authority (CA) bundle.
}
\examples{
library(rpki)
configure_download_file_pki() # will prompt for passphrase
download.file("https://httpbin.org/")
install.packages("my_private_package")

library(rpki)
configure_download_file_pki(password = "my_pki_passphrase")
GET("http://httpbin.org/")
install.packages("my_private_package")
}
