% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config_pki.R
\name{auto_config_pki}
\alias{auto_config_pki}
\title{Set httr config parameters}
\usage{
auto_config_pki(password = NULL)
}
\arguments{
\item{password}{string: passphrase used to encrypt/decrypt the private key of a PKI certificate}
}
\description{
Override configuration settings used by the 'httr' package to allow working with PKI-enabled web services.
}
\details{
Default mypki configuration settings will be used. If no password is provided (\code{password = NULL}), the PKI certificate specified by the mypki configuration file is assumed to be unencrypted. The PKI certificate file must be in PKCS#12 format. The following 'httr' config settings get modified: cainfo, sslcert, sslkey.

If a mypki configuration file cannot be found, users are prompted for file paths to both a PKI certificate and a Certificate Authority (CA) bundle.
}
\examples{
\dontrun{
library(rpki)
auto_config_pki()
GET('https://your.pki.enabled.website/path/to/whatever')
}

}
