% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config_pki.R
\name{auto_config_pki}
\alias{auto_config_pki}
\title{Set httr config parameters}
\usage{
auto_config_pki(password = NULL)
}
\arguments{
\item{password}{string: passphrase used to encrypt the private key of p12_file. Optional argument that defaults to NULL.}
}
\description{
A wrapper function that automatically overrides configuration settings used by the httr package to allow working with PKI-enabled web services.
}
\details{
This function should be called for automatic configuration. The PKI certificate and private key are extracted from a PKCS#12-formatted PKI file and used to define the following httr config settings: cainfo, sslcert, sslkey
}
\examples{
library(rpki)
auto_config_pki()
GET('https://your.pki.enabled.website/path/to/whatever')

}
