% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_config_pki.R
\name{manual_config_pki}
\alias{manual_config_pki}
\title{Set httr config parameters}
\usage{
manual_config_pki(pki_file = NULL, ca_file = NULL, password = NULL,
  mypki_file = NULL)
}
\arguments{
\item{pki_file}{string: absolute file path to a pki certificate}

\item{ca_file}{string: absolute file path to a Certificate Authority (CA) bundle}

\item{password}{string: passphrase used to encrypt/decrypt the private key of the pki certificate}

\item{mypki_file}{string: absolute file path to save mypki configuration file. Defaults to the home directory}
}
\description{
Override configuration settings used by the 'httr' package to allow working with PKI-enabled web services, based on the supplied arguments.
}
\details{
This function should only be used when default arguments are not desired using \code{auto_config_pki()}. A mypki file is generated based on the file paths of a PKI certificate and Certificate Authority bundle. At a minimum, \code{pki_file} and \code{ca_file} must be specified. If \code{mypki_file} is not specified, a mypki configuration file is generated in the user's home directory. The PKI certificate must be in PKCS#12 format. The following 'httr' config settings get modified: cainfo, sslcert, sslkey
}
\examples{
\dontrun{
library(rpki)
manual_config_pki(pki_file = 'path/to/my/pki.p12',
                  ca_file = '/path/to/certificate_authority.crt',
                  password = 'my_pki_passphrase')
GET('https://your.pki.enabled.website/path/to/whatever')
}

}
