% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki_enable_download_file.R
\name{pki_enable_download_file}
\alias{pki_enable_download_file}
\title{Set download.file parameters}
\usage{
pki_enable_download_file(mypki_file = NULL, pki_file = NULL,
  ca_file = NULL, password = NULL, overwrite = FALSE)
}
\arguments{
\item{mypki_file}{string: absolute file path to save .mypki configuration file. Defaults to the home directory}

\item{pki_file}{string: absolute file path to a pki certificate}

\item{ca_file}{string: absolute file path to a Certificate Authority (CA) bundle}

\item{password}{string: passphrase used to encrypt/decrypt the PKI certificate}

\item{overwrite}{logical: overwrite a pre-existing mypki configuration file, if found. Defaults to FALSE}
}
\description{
Override default configuration settings used by \code{download.file()} to allow for PKI-enabled use of \code{install.packages()}.
}
\details{
All arguments are optional. Default mypki configuration settings will be used unless otherwise specified. The PKI certificate file must be in PKCS#12 format. The arguments of \code{download.file()} are modified so that \code{method = "curl"} and the following extra curl command-line arguments are set: cacert, cert, key, pass. See \code{?download.file} for more information.

This download method requires curl to be installed and added to the PATH environment variable. This is already done for most Linux/Mac OS's.

If a mypki configuration file cannot be found, users are prompted for file paths to both a PKI certificate and a Certificate Authority (CA) bundle.
}
\examples{
library(rpki)
pki_enable_download_file() # will prompt for passphrase
download.file("https://httpbin.org/")
install.packages("my_private_package")

library(rpki)
pki_enable_download_file(password = "my_pki_passphrase")
GET("http://httpbin.org/")
install.packages("my_private_package")
}
